==> Microbiota/2_DADA2_pipeline_all.Rmd <==
---
title: "DADA2 pipeline"
output:
  github_document:
    toc: true
editor_options:
  chunk_output_type: console
---

Read in packages:
```{r}
library(ggplot2)
library(dplyr)
library(gridExtra)
library(dada2)
library(phyloseq)
library(phangorn)

packageVersion("dada2")

meta_data <- read.csv(file = "~/vol1/all/meta_data_prepared.txt", sep="\t")
```

Prepare:
```{r}
#Withdraw the projects
projects <- unique(meta_data$BioProject)

#Exclude: PRJNA606913
projects <- projects[projects != "PRJNA606913"]
projects <- projects[!is.na(projects)] 

#Cutadapt projects (they will be in a subfolder)
cutadapt_projects <- c("PRJNA1013140", "PRJNA799831", "PRJNA800628", "PRJNA847791", "PRJNA850658", "PRJNA892439")
```


Read in library with fastq-files:
```{r}
for (i in setdiff(projects, cutadapt_projects)){
  assign(paste0("path_", i), paste0("~/vol1/all/", i))
}

for (i in cutadapt_projects){
  assign(paste0("path_", i), paste0("~/vol1/all/", i, "/cutadapt"))
}
```

Filter and trim:
```{r}

# Sort ensures forward/reverse reads are in same order
for (i in projects){
  assign(paste0("fnFs_", i),  sort(list.files(eval(parse(text=paste0("path_", i))), pattern="_1.fastq.gz", full.names = T)))
  assign(paste0("fnRs_", i),  sort(list.files(eval(parse(text=paste0("path_", i))), pattern="_2.fastq.gz", full.names = T)))
}


# Extract sample names, assuming filenames have format: SAMPLENAME_XXX.fastq
for (i in projects){
  assign(paste0("sample.names_", i), sapply(strsplit(basename(eval(parse(text=paste0("fnFs_", i)))), "_"), `[`, 1))
}


# See quality: 
for (i in setdiff(projects, cutadapt_projects)){
  png(paste0("~/vol1/all/Forward_QC_", i, ".tiff"))
  print(plotQualityProfile(eval(parse(text=paste0("fnFs_", i, "[1:2]")))))
  dev.off()
  
  png(paste0("~/vol1/all/Reverse_QC_", i, ".tiff"))
  print(plotQualityProfile(eval(parse(text=paste0("fnRs_", i, "[1:2]")))))
  dev.off()
}

```

Prepare for trim ends: 
```{r}
for (i in projects){
  assign(paste0("filtFs_", i), file.path(eval(parse(text=paste0("path_", i))), "filtered", paste0(eval(parse(text=paste0("sample.names_", i))), "_F_filt.fastq.gz")))
  assign(paste0("filtRs_", i), file.path(eval(parse(text=paste0("path_", i))), "filtered", paste0(eval(parse(text=paste0("sample.names_", i))), "_R_filt.fastq.gz")))
}


names(filtFs_PRJEB21819) <- sample.names_PRJEB21819
names(filtRs_PRJEB21819) <- sample.names_PRJEB21819
names(filtFs_PRJNA1024059) <- sample.names_PRJNA1024059
names(filtRs_PRJNA1024059) <- sample.names_PRJNA1024059
names(filtFs_PRJNA1029396) <- sample.names_PRJNA1029396
names(filtRs_PRJNA1029396) <- sample.names_PRJNA1029396
names(filtFs_PRJNA1053658) <- sample.names_PRJNA1053658
names(filtRs_PRJNA1053658) <- sample.names_PRJNA1053658
names(filtFs_PRJNA1156939) <- sample.names_PRJNA1156939
names(filtRs_PRJNA1156939) <- sample.names_PRJNA1156939
names(filtFs_PRJNA414072) <- sample.names_PRJNA414072
names(filtRs_PRJNA414072) <- sample.names_PRJNA414072
names(filtFs_PRJNA511932) <- sample.names_PRJNA511932
names(filtRs_PRJNA511932) <- sample.names_PRJNA511932
names(filtFs_PRJNA515212) <- sample.names_PRJNA515212
names(filtRs_PRJNA515212) <- sample.names_PRJNA515212
names(filtFs_PRJNA565903) <- sample.names_PRJNA565903
names(filtRs_PRJNA565903) <- sample.names_PRJNA565903
names(filtFs_PRJNA574485) <- sample.names_PRJNA574485
names(filtRs_PRJNA574485) <- sample.names_PRJNA574485
names(filtFs_PRJNA574565) <- sample.names_PRJNA574565
names(filtRs_PRJNA574565) <- sample.names_PRJNA574565
names(filtFs_PRJNA628832) <- sample.names_PRJNA628832
names(filtRs_PRJNA628832) <- sample.names_PRJNA628832
names(filtFs_PRJNA633233) <- sample.names_PRJNA633233
names(filtRs_PRJNA633233) <- sample.names_PRJNA633233
names(filtFs_PRJNA633959) <- sample.names_PRJNA633959
names(filtRs_PRJNA633959) <- sample.names_PRJNA633959
names(filtFs_PRJNA646271) <- sample.names_PRJNA646271
names(filtRs_PRJNA646271) <- sample.names_PRJNA646271
names(filtFs_PRJNA646468) <- sample.names_PRJNA646468
names(filtRs_PRJNA646468) <- sample.names_PRJNA646468
names(filtFs_PRJNA679275) <- sample.names_PRJNA679275
names(filtRs_PRJNA679275) <- sample.names_PRJNA679275
names(filtFs_PRJNA680445) <- sample.names_PRJNA680445
names(filtRs_PRJNA680445) <- sample.names_PRJNA680445
names(filtFs_PRJNA681685) <- sample.names_PRJNA681685
names(filtRs_PRJNA681685) <- sample.names_PRJNA681685
names(filtFs_PRJNA695309) <- sample.names_PRJNA695309
names(filtRs_PRJNA695309) <- sample.names_PRJNA695309
names(filtFs_PRJNA732664) <- sample.names_PRJNA732664
names(filtRs_PRJNA732664) <- sample.names_PRJNA732664
names(filtFs_PRJNA753264) <- sample.names_PRJNA753264
names(filtRs_PRJNA753264) <- sample.names_PRJNA753264
names(filtFs_PRJNA799831) <- sample.names_PRJNA799831
names(filtRs_PRJNA799831) <- sample.names_PRJNA799831
names(filtFs_PRJNA850658) <- sample.names_PRJNA850658
names(filtRs_PRJNA850658) <- sample.names_PRJNA850658
names(filtFs_PRJNA890948) <- sample.names_PRJNA890948
names(filtRs_PRJNA890948) <- sample.names_PRJNA890948
names(filtFs_PRJNA892439) <- sample.names_PRJNA892439
names(filtRs_PRJNA892439) <- sample.names_PRJNA892439
names(filtFs_PRJNA917086) <- sample.names_PRJNA917086
names(filtRs_PRJNA917086) <- sample.names_PRJNA917086
names(filtFs_PRJNA935155) <- sample.names_PRJNA935155
names(filtRs_PRJNA935155) <- sample.names_PRJNA935155
names(filtFs_PRJNA938107) <- sample.names_PRJNA938107
names(filtRs_PRJNA938107) <- sample.names_PRJNA938107
names(filtFs_PRJNA950106) <- sample.names_PRJNA950106
names(filtRs_PRJNA950106) <- sample.names_PRJNA950106
names(filtFs_PRJNA806451) <- sample.names_PRJNA806451
names(filtRs_PRJNA806451) <- sample.names_PRJNA806451
names(filtFs_PRJNA804422) <- sample.names_PRJNA804422
names(filtRs_PRJNA804422) <- sample.names_PRJNA804422
names(filtFs_PRJNA800628) <- sample.names_PRJNA800628
names(filtRs_PRJNA800628) <- sample.names_PRJNA800628
names(filtFs_PRJNA1013140) <- sample.names_PRJNA1013140
names(filtRs_PRJNA1013140) <- sample.names_PRJNA1013140
names(filtFs_PRJNA961636) <- sample.names_PRJNA961636
names(filtRs_PRJNA961636) <- sample.names_PRJNA961636
names(filtFs_PRJNA871997) <- sample.names_PRJNA871997
names(filtRs_PRJNA871997) <- sample.names_PRJNA871997
names(filtFs_PRJNA847791) <- sample.names_PRJNA847791
names(filtRs_PRJNA847791) <- sample.names_PRJNA847791
names(filtFs_PRJNA798680) <- sample.names_PRJNA798680
names(filtRs_PRJNA798680) <- sample.names_PRJNA798680
names(filtFs_PRJEB67783) <- sample.names_PRJEB67783
names(filtRs_PRJEB67783) <- sample.names_PRJEB67783
names(filtFs_PRJNA867125) <- sample.names_PRJNA867125
names(filtRs_PRJNA867125) <- sample.names_PRJNA867125
names(filtFs_PRJNA812940) <- sample.names_PRJNA812940
names(filtRs_PRJNA812940) <- sample.names_PRJNA812940
names(filtFs_PRJNA1078442) <- sample.names_PRJNA1078442
names(filtRs_PRJNA1078442) <- sample.names_PRJNA1078442
names(filtFs_PRJNA877434) <- sample.names_PRJNA877434
names(filtRs_PRJNA877434) <- sample.names_PRJNA877434
names(filtFs_PRJNA832901) <- sample.names_PRJNA832901
names(filtRs_PRJNA832901) <- sample.names_PRJNA832901
names(filtFs_PRJEB41297) <- sample.names_PRJEB41297
names(filtRs_PRJEB41297) <- sample.names_PRJEB41297
names(filtFs_PRJNA660322) <- sample.names_PRJNA660322
names(filtRs_PRJNA660322) <- sample.names_PRJNA660322
names(filtFs_PRJNA1101190) <- sample.names_PRJNA1101190
names(filtRs_PRJNA1101190) <- sample.names_PRJNA1101190
names(filtFs_PRJNA904086) <- sample.names_PRJNA904086
names(filtRs_PRJNA904086) <- sample.names_PRJNA904086
names(filtFs_PRJNA883162) <- sample.names_PRJNA883162
names(filtRs_PRJNA883162) <- sample.names_PRJNA883162
names(filtFs_PRJEB52971) <- sample.names_PRJEB52971
names(filtRs_PRJEB52971) <- sample.names_PRJEB52971
names(filtFs_PRJNA761343) <- sample.names_PRJNA761343
names(filtRs_PRJNA761343) <- sample.names_PRJNA761343
names(filtFs_PRJEB39044) <- sample.names_PRJEB39044
names(filtRs_PRJEB39044) <- sample.names_PRJEB39044
names(filtFs_PRJNA1084087) <- sample.names_PRJNA1084087
names(filtRs_PRJNA1084087) <- sample.names_PRJNA1084087
names(filtFs_PRJNA791216) <- sample.names_PRJNA791216
names(filtRs_PRJNA791216) <- sample.names_PRJNA791216
names(filtFs_PRJNA789833) <- sample.names_PRJNA789833
names(filtRs_PRJNA789833) <- sample.names_PRJNA789833
names(filtFs_PRJNA786110) <- sample.names_PRJNA786110
names(filtRs_PRJNA786110) <- sample.names_PRJNA786110
names(filtFs_PRJNA759926) <- sample.names_PRJNA759926
names(filtRs_PRJNA759926) <- sample.names_PRJNA759926
names(filtFs_PRJEB49838) <- sample.names_PRJEB49838
names(filtRs_PRJEB49838) <- sample.names_PRJEB49838
names(filtFs_PRJNA955157) <- sample.names_PRJNA955157
names(filtRs_PRJNA955157) <- sample.names_PRJNA955157
names(filtFs_PRJNA493726) <- sample.names_PRJNA493726
names(filtRs_PRJNA493726) <- sample.names_PRJNA493726

```

Trim ends: 
-trimLeft according to alignment 
-truncLen according to quality 
-maxEE can be changed if not enough passes filter
-Lenght = 410. Do not trim too small!
```{r}

out_PRJNA679275 <- filterAndTrim(fnFs_PRJNA679275, filtFs_PRJNA679275, fnRs_PRJNA679275, filtRs_PRJNA679275, truncLen=c(250,230), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA679275)
out_PRJNA646271 <- filterAndTrim(fnFs_PRJNA646271, filtFs_PRJNA646271, fnRs_PRJNA646271, filtRs_PRJNA646271, truncLen=c(265,205), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA646271)
out_PRJNA681685 <- filterAndTrim(fnFs_PRJNA681685, filtFs_PRJNA681685, fnRs_PRJNA681685, filtRs_PRJNA681685, truncLen=c(250,250), trimLeft=c(33, 39), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA681685)
out_PRJNA565903 <- filterAndTrim(fnFs_PRJNA565903, filtFs_PRJNA565903, fnRs_PRJNA565903, filtRs_PRJNA565903, truncLen=c(250,230), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA565903)
out_PRJNA515212 <- filterAndTrim(fnFs_PRJNA515212, filtFs_PRJNA515212, fnRs_PRJNA515212, filtRs_PRJNA515212, truncLen=c(240,240), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA515212)
out_PRJEB21819 <- filterAndTrim(fnFs_PRJEB21819, filtFs_PRJEB21819, fnRs_PRJEB21819, filtRs_PRJEB21819, truncLen=c(250,230), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJEB21819)
out_PRJNA414072 <- filterAndTrim(fnFs_PRJNA414072, filtFs_PRJNA414072, fnRs_PRJNA414072, filtRs_PRJNA414072, truncLen=c(245,225), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA414072)
out_PRJNA753264 <- filterAndTrim(fnFs_PRJNA753264, filtFs_PRJNA753264, fnRs_PRJNA753264, filtRs_PRJNA753264, truncLen=c(240,240), trimLeft=c(29, 35), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA753264) #(v)
out_PRJNA574565 <- filterAndTrim(fnFs_PRJNA574565, filtFs_PRJNA574565, fnRs_PRJNA574565, filtRs_PRJNA574565, truncLen=c(260,220), trimLeft=c(33, 39), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA574565)
out_PRJNA732664 <- filterAndTrim(fnFs_PRJNA732664, filtFs_PRJNA732664, fnRs_PRJNA732664, filtRs_PRJNA732664, truncLen=c(250,240), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA732664)
out_PRJNA574485 <- filterAndTrim(fnFs_PRJNA574485, filtFs_PRJNA574485, fnRs_PRJNA574485, filtRs_PRJNA574485, truncLen=c(250,220), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA574485)
out_PRJNA633959 <- filterAndTrim(fnFs_PRJNA633959, filtFs_PRJNA633959, fnRs_PRJNA633959, filtRs_PRJNA633959, truncLen=c(255,225), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA633959)
out_PRJNA633233 <- filterAndTrim(fnFs_PRJNA633233, filtFs_PRJNA633233, fnRs_PRJNA633233, filtRs_PRJNA633233, truncLen=c(240,240), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA633233)
out_PRJNA628832 <- filterAndTrim(fnFs_PRJNA628832, filtFs_PRJNA628832, fnRs_PRJNA628832, filtRs_PRJNA628832, truncLen=c(245,230), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA628832)
out_PRJNA680445 <- filterAndTrim(fnFs_PRJNA680445, filtFs_PRJNA680445, fnRs_PRJNA680445, filtRs_PRJNA680445, truncLen=c(250,230), trimLeft=c(24, 13), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA680445) #(v)
out_PRJNA511932 <- filterAndTrim(fnFs_PRJNA511932, filtFs_PRJNA511932, fnRs_PRJNA511932, filtRs_PRJNA511932, truncLen=c(245,235), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA511932)
out_PRJNA646468 <- filterAndTrim(fnFs_PRJNA646468, filtFs_PRJNA646468, fnRs_PRJNA646468, filtRs_PRJNA646468, truncLen=c(245,240), trimLeft=c(27, 33), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA646468)
out_PRJNA1024059 <- filterAndTrim(fnFs_PRJNA1024059, filtFs_PRJNA1024059, fnRs_PRJNA1024059, filtRs_PRJNA1024059, truncLen=c(250,230), trimLeft=c(33, 39), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA1024059)
out_PRJNA1029396 <- filterAndTrim(fnFs_PRJNA1029396, filtFs_PRJNA1029396, fnRs_PRJNA1029396, filtRs_PRJNA1029396, truncLen=c(240,230), trimLeft=c(26, 33), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA1029396)
out_PRJNA1053658 <- filterAndTrim(fnFs_PRJNA1053658, filtFs_PRJNA1053658, fnRs_PRJNA1053658, filtRs_PRJNA1053658, truncLen=c(250,230), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA1053658)
out_PRJNA1156939 <- filterAndTrim(fnFs_PRJNA1156939, filtFs_PRJNA1156939, fnRs_PRJNA1156939, filtRs_PRJNA1156939, truncLen=c(245,235), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA1156939)
out_PRJNA695309 <- filterAndTrim(fnFs_PRJNA695309, filtFs_PRJNA695309, fnRs_PRJNA695309, filtRs_PRJNA695309, truncLen=c(245,235), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA695309)
out_PRJNA799831 <- filterAndTrim(fnFs_PRJNA799831, filtFs_PRJNA799831, fnRs_PRJNA799831, filtRs_PRJNA799831, truncLen=c(225,225), trimLeft=c(19, 1), maxN=0, maxEE=c(20,20), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA799831) #Poor quality
out_PRJNA850658 <- filterAndTrim(fnFs_PRJNA850658, filtFs_PRJNA850658, fnRs_PRJNA850658, filtRs_PRJNA850658, truncLen=c(230,225), trimLeft=c(20, 1), maxN=0, maxEE=c(7,7), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA850658) #(v)
out_PRJNA890948 <- filterAndTrim(fnFs_PRJNA890948, filtFs_PRJNA890948, fnRs_PRJNA890948, filtRs_PRJNA890948, truncLen=c(240,240), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA890948)
out_PRJNA892439 <- filterAndTrim(fnFs_PRJNA892439, filtFs_PRJNA892439, fnRs_PRJNA892439, filtRs_PRJNA892439, truncLen=c(225,225), trimLeft=c(20, 39), maxN=0, maxEE=c(20,20), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA892439) #Poor quality!
out_PRJNA917086 <- filterAndTrim(fnFs_PRJNA917086, filtFs_PRJNA917086, fnRs_PRJNA917086, filtRs_PRJNA917086, truncLen=c(250,230), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA917086)
out_PRJNA935155 <- filterAndTrim(fnFs_PRJNA935155, filtFs_PRJNA935155, fnRs_PRJNA935155, filtRs_PRJNA935155, truncLen=c(250,225), trimLeft=c(13, 48), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA935155) #(v)
out_PRJNA938107 <- filterAndTrim(fnFs_PRJNA938107, filtFs_PRJNA938107, fnRs_PRJNA938107, filtRs_PRJNA938107, truncLen=c(240,240), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA938107)
#out_PRJNA950106 <- filterAndTrim(fnFs_PRJNA950106, filtFs_PRJNA950106, fnRs_PRJNA950106, filtRs_PRJNA950106, truncLen=c(150,150), trimLeft=c(0, 13), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T) -> too short!
#head(out_PRJNA950106)
out_PRJNA806451 <- filterAndTrim(fnFs_PRJNA806451, filtFs_PRJNA806451, fnRs_PRJNA806451, filtRs_PRJNA806451, truncLen=c(240,240), trimLeft=c(33, 39), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA806451)
out_PRJNA804422 <- filterAndTrim(fnFs_PRJNA804422, filtFs_PRJNA804422, fnRs_PRJNA804422, filtRs_PRJNA804422, truncLen=c(245,230), trimLeft=c(28, 33), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA804422)
out_PRJNA800628 <- filterAndTrim(fnFs_PRJNA800628, filtFs_PRJNA800628, fnRs_PRJNA800628, filtRs_PRJNA800628, truncLen=c(225,225), trimLeft=c(4, 6), maxN=0, maxEE=c(20,20), truncQ=2, rm.phix=T, compress=T, multithread=T) #Not good enough quality!
head(out_PRJNA800628)
out_PRJNA1013140 <- filterAndTrim(fnFs_PRJNA1013140, filtFs_PRJNA1013140, fnRs_PRJNA1013140, filtRs_PRJNA1013140, truncLen=c(235,225), trimLeft=c(19, 1), maxN=0, maxEE=c(10,10), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA1013140)
out_PRJNA961636 <- filterAndTrim(fnFs_PRJNA961636, filtFs_PRJNA961636, fnRs_PRJNA961636, filtRs_PRJNA961636, truncLen=c(250,230), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA961636)
out_PRJNA871997 <- filterAndTrim(fnFs_PRJNA871997, filtFs_PRJNA871997, fnRs_PRJNA871997, filtRs_PRJNA871997, truncLen=c(250,230), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA871997)
out_PRJNA847791 <- filterAndTrim(fnFs_PRJNA847791, filtFs_PRJNA847791, fnRs_PRJNA847791, filtRs_PRJNA847791, truncLen=c(235,215), trimLeft=c(20, 1), maxN=0, maxEE=c(10,10), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA847791) #(v)
out_PRJNA798680 <- filterAndTrim(fnFs_PRJNA798680, filtFs_PRJNA798680, fnRs_PRJNA798680, filtRs_PRJNA798680, truncLen=c(250,230), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA798680)
out_PRJEB67783 <- filterAndTrim(fnFs_PRJEB67783, filtFs_PRJEB67783, fnRs_PRJEB67783, filtRs_PRJEB67783, truncLen=c(250,230), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJEB67783)
out_PRJNA867125 <- filterAndTrim(fnFs_PRJNA867125, filtFs_PRJNA867125, fnRs_PRJNA867125, filtRs_PRJNA867125, truncLen=c(240,240), trimLeft=c(26, 39), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA867125)
out_PRJNA812940 <- filterAndTrim(fnFs_PRJNA812940, filtFs_PRJNA812940, fnRs_PRJNA812940, filtRs_PRJNA812940, truncLen=c(250,230), trimLeft=c(37, 38), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA812940) #(v)
out_PRJNA1078442 <- filterAndTrim(fnFs_PRJNA1078442, filtFs_PRJNA1078442, fnRs_PRJNA1078442, filtRs_PRJNA1078442, truncLen=c(240,240), trimLeft=c(30, 38), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA1078442)
out_PRJNA877434 <- filterAndTrim(fnFs_PRJNA877434, filtFs_PRJNA877434, fnRs_PRJNA877434, filtRs_PRJNA877434, truncLen=c(240,240), trimLeft=c(40, 47), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA877434)
out_PRJNA832901 <- filterAndTrim(fnFs_PRJNA832901, filtFs_PRJNA832901, fnRs_PRJNA832901, filtRs_PRJNA832901, truncLen=c(245,235), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA832901)
out_PRJEB41297 <- filterAndTrim(fnFs_PRJEB41297, filtFs_PRJEB41297, fnRs_PRJEB41297, filtRs_PRJEB41297, truncLen=c(245,235), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJEB41297)
out_PRJNA660322 <- filterAndTrim(fnFs_PRJNA660322, filtFs_PRJNA660322, fnRs_PRJNA660322, filtRs_PRJNA660322, truncLen=c(245,235), trimLeft=c(33, 39), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA660322)
out_PRJNA1101190 <- filterAndTrim(fnFs_PRJNA1101190, filtFs_PRJNA1101190, fnRs_PRJNA1101190, filtRs_PRJNA1101190, truncLen=c(250,230), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA1101190)
#out_PRJNA904086 <- filterAndTrim(fnFs_PRJNA904086, filtFs_PRJNA904086, fnRs_PRJNA904086, filtRs_PRJNA904086, truncLen=c(240,240), trimLeft=c(0, 33), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T) --> too short!
#head(out_PRJNA904086)
out_PRJNA883162 <- filterAndTrim(fnFs_PRJNA883162, filtFs_PRJNA883162, fnRs_PRJNA883162, filtRs_PRJNA883162, truncLen=c(240,240), trimLeft=c(27, 33), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA883162)
out_PRJEB52971 <- filterAndTrim(fnFs_PRJEB52971, filtFs_PRJEB52971, fnRs_PRJEB52971, filtRs_PRJEB52971, truncLen=c(230,250), trimLeft=c(24, 31), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJEB52971)
out_PRJNA761343 <- filterAndTrim(fnFs_PRJNA761343, filtFs_PRJNA761343, fnRs_PRJNA761343, filtRs_PRJNA761343, truncLen=c(245,235), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA761343)
out_PRJEB39044 <- filterAndTrim(fnFs_PRJEB39044, filtFs_PRJEB39044, fnRs_PRJEB39044, filtRs_PRJEB39044, truncLen=c(245,235), trimLeft=c(27, 33), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJEB39044)
out_PRJNA1084087 <- filterAndTrim(fnFs_PRJNA1084087, filtFs_PRJNA1084087, fnRs_PRJNA1084087, filtRs_PRJNA1084087, truncLen=c(240,240), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA1084087)
out_PRJNA791216 <- filterAndTrim(fnFs_PRJNA791216, filtFs_PRJNA791216, fnRs_PRJNA791216, filtRs_PRJNA791216, truncLen=c(245,225), trimLeft=c(14, 38), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA791216)
out_PRJNA789833 <- filterAndTrim(fnFs_PRJNA789833, filtFs_PRJNA789833, fnRs_PRJNA789833, filtRs_PRJNA789833, truncLen=c(260,220), trimLeft=c(36, 41), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA789833)
out_PRJNA786110 <- filterAndTrim(fnFs_PRJNA786110, filtFs_PRJNA786110, fnRs_PRJNA786110, filtRs_PRJNA786110, truncLen=c(240,240), trimLeft=c(24, 33), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA786110)
out_PRJNA759926 <- filterAndTrim(fnFs_PRJNA759926, filtFs_PRJNA759926, fnRs_PRJNA759926, filtRs_PRJNA759926, truncLen=c(250,230), trimLeft=c(24, 32), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA759926)
out_PRJEB49838 <- filterAndTrim(fnFs_PRJEB49838, filtFs_PRJEB49838, fnRs_PRJEB49838, filtRs_PRJEB49838, truncLen=c(245,235), trimLeft=c(18, 21), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJEB49838)
out_PRJNA955157 <- filterAndTrim(fnFs_PRJNA955157, filtFs_PRJNA955157, fnRs_PRJNA955157, filtRs_PRJNA955157, truncLen=c(240,240), trimLeft=c(28, 33), maxN=0, maxEE=c(5,5), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA955157)
out_PRJNA493726 <- filterAndTrim(fnFs_PRJNA493726, filtFs_PRJNA493726, fnRs_PRJNA493726, filtRs_PRJNA493726, truncLen=c(220,210), trimLeft=c(0, 10), maxN=0, maxEE=c(20,20), truncQ=2, rm.phix=T, compress=T, multithread=T)
head(out_PRJNA493726) #Too poor quality
```

Learn error rates:
```{r}
#Exclude PRJNA950106, PRJNA904086 --> too short
projects <- projects[!(projects %in% c("PRJNA950106", "PRJNA904086"))]

#Exclude with poor quality:
projects <- projects[!(projects %in% c("PRJNA493726", "PRJNA800628", "PRJNA892439"))]

#For paired-end:
for (i in projects){
  assign(paste0("errF_", i), learnErrors(eval(parse(text=paste0("filtFs_", i))), multithread=T))
  assign(paste0("errR_", i), learnErrors(eval(parse(text=paste0("filtRs_", i))), multithread=T))
}

```

Sample inference:
```{r}

# Sample inference and merger of paired-end reads, Illumina
for (i in projects[3:55]){ 
  mergers <- vector("list", length(eval(parse(text=paste0("sample.names_", i)))))
  names(mergers) <- eval(parse(text=paste0("sample.names_", i)))
for(sam in eval(parse(text=paste0("sample.names_", i)))) {
  cat("Processing:", sam, "\n")
    derepF <- derepFastq(eval(parse(text=paste0("filtFs_", i)))[[sam]])
    ddF <- dada(derepF, err=eval(parse(text=paste0("errF_", i))), multithread=TRUE)
    derepR <- derepFastq(eval(parse(text=paste0("filtRs_", i)))[[sam]])
    ddR <- dada(derepR, err=eval(parse(text=paste0("errR_", i))), multithread=TRUE)
    merger <- mergePairs(ddF, derepF, ddR, derepR)
    mergers[[sam]] <- merger
}
rm(derepF); rm(derepR)
# Construct sequence table and remove chimeras
seqtab <- makeSequenceTable(mergers)
saveRDS(seqtab, paste0("~/vol1/all/", i, "/seqtab.rds"))
rm(mergers, seqtab)s
}

getN <- function(x) sum(getUniques(x))

#Check progress:
for (i in projects){
  seqtab <- readRDS(paste0("~/vol1/all/", i, "/seqtab.rds"))
  track <- cbind(eval(parse(text=paste0("out_", i))), rowSums(seqtab))
  print(i)
  print(head(track))
  write.table(track, file=paste0("~/vol1/all/", i, "/DADA2_QC.txt"))
  rm(seqtab, track)
} 

```

Merge dataframes:
```{r}
# Merge multiple runs (if necessary)
for (i in projects){
  assign(paste0("st_", i), readRDS(paste0("~/vol1/all/", i, "/seqtab.rds")))
}

st.all <- mergeSequenceTables(st_PRJEB21819, st_PRJNA1024059, st_PRJNA1029396, st_PRJNA1053658, st_PRJNA1156939, st_PRJNA414072, st_PRJNA511932, st_PRJNA515212, st_PRJNA565903, st_PRJNA574485, st_PRJNA574565, st_PRJNA628832, st_PRJNA633233, st_PRJNA633959, st_PRJNA646271, st_PRJNA646468, st_PRJNA679275, st_PRJNA680445, st_PRJNA681685, st_PRJNA695309, st_PRJNA732664, st_PRJNA753264, st_PRJNA799831, st_PRJNA850658, st_PRJNA890948, st_PRJNA917086, st_PRJNA935155, st_PRJNA938107, st_PRJNA806451, st_PRJNA804422, st_PRJNA1013140, st_PRJNA961636, st_PRJNA871997, st_PRJNA847791, st_PRJNA798680, st_PRJEB67783, st_PRJNA867125, st_PRJNA812940, st_PRJNA1078442, st_PRJNA877434, st_PRJNA832901, st_PRJEB41297, st_PRJNA660322, st_PRJNA1101190, st_PRJNA883162, st_PRJEB52971, st_PRJNA761343, st_PRJEB39044, st_PRJNA1084087, st_PRJNA791216, st_PRJNA789833, st_PRJNA786110, st_PRJNA759926, st_PRJEB49838, st_PRJNA955157)


rm(st_PRJEB21819, st_PRJNA1024059, st_PRJNA1029396, st_PRJNA1053658, st_PRJNA1156939, st_PRJNA414072, st_PRJNA511932, st_PRJNA515212, st_PRJNA565903, st_PRJNA574485, st_PRJNA574565, st_PRJNA628832, st_PRJNA633233, st_PRJNA633959, st_PRJNA646271, st_PRJNA646468, st_PRJNA679275, st_PRJNA680445, st_PRJNA681685, st_PRJNA695309, st_PRJNA732664, st_PRJNA753264, st_PRJNA799831, st_PRJNA850658, st_PRJNA890948, st_PRJNA917086, st_PRJNA935155, st_PRJNA938107, st_PRJNA806451, st_PRJNA804422, st_PRJNA1013140, st_PRJNA961636, st_PRJNA871997, st_PRJNA847791, st_PRJNA798680, st_PRJEB67783, st_PRJNA867125, st_PRJNA812940, st_PRJNA1078442, st_PRJNA877434, st_PRJNA832901, st_PRJEB41297, st_PRJNA660322, st_PRJNA1101190, st_PRJNA883162, st_PRJEB52971, st_PRJNA761343, st_PRJEB39044, st_PRJNA1084087, st_PRJNA791216, st_PRJNA789833, st_PRJNA786110, st_PRJNA759926, st_PRJEB49838, st_PRJNA955157)

saveRDS(st.all, "~/vol1/all/Total_seqtab.rds")
```


Remove chimeras:
```{r}
seqtab <- removeBimeraDenovo(st.all, method="consensus", multithread=TRUE)
saveRDS(seqtab, "~/vol1/all/Total_seqtab_nochim.rds")

sum(seqtab)/sum(st.all)
rm(st.all)
```

Assign taxonomy and construct phyloseq object:
```{r}
#Genus level:
taxa <- assignTaxonomy(seqtab, "~/vol1/silva_nr99_v138.1_train_set.fa.gz", multithread=T)

#Add species level
#taxa_extension <- addSpecies(taxa, "./vol1/silva_species_assignment_v138.1.fa.gz")

#Make phyloseq object:
sam <- meta_data
row.names(sam) <- sam$Run
ps <- phyloseq(otu_table(seqtab, taxa_are_rows=F), sample_data(sam), tax_table(taxa_extension))

#Shorter names for ASVs:
dna <- Biostrings::DNAStringSet(taxa_names(ps))
names(dna) <- taxa_names(ps)
ps <- merge_phyloseq(ps, dna)
taxa_names(ps) <- paste0("ASV", seq(ntaxa(ps)))

#Transforn into proportions:
ps.prop <- transform_sample_counts(ps, function(otu) otu/sum(otu))

#Save!:
saveRDS(ps, file = paste0("~/vol1/all/", i, "/ps_count.rds"))
saveRDS(ps.prop, file = paste0("~/vol1/all/", i, "/ps_relab.rds"))

```
