==> Microbiota/3_Prepare_microbiome_data.Rmd <==
---
title: "QC and preparation of data"
output:
  github_document:
    toc: true
editor_options:
  chunk_output_type: console
---


## Read in packages and data
```{r message=FALSE, warning=FALSE, error=TRUE}
library(dplyr)
library(tidyverse)
library(ggplot2)
library(phyloseq)
library(vegan)
library(ggpubr)

packageVersion("dplyr")
packageVersion("tidyverse")
packageVersion("ggplot2")
packageVersion("phyloseq")
packageVersion("vegan")
packageVersion("ggpubr")

path <- "C:/Users/FX76TZ/OneDrive - Aalborg Universitet/Disease correlations/"
path <- "~/vol1/all/"
```


Read in all ps objects:
```{r}
meta_data <- read.csv(file = "~/vol1/all/meta_data_prepared.txt", sep="\t")
head(meta_data)

ps <- readRDS(paste0(path, "ps_count.rds"))

ps #4005 samples, 153773 taxa
head(sample_data(ps))
otu_table(ps)[1:7, 1:7] #Taxa are columns, samples are rows
tax_table(ps)[1:7, 1:6]

table(sample_data(ps)$Disease)

```

## QC based on read depth
```{r cache=TRUE}

p <- ggplot()+
  geom_histogram(mapping=aes(sample_sums(ps)), fill = "lightskyblue1", color="black", bins = 100)+
  theme_classic()+xlab("Read depth")+ylab("Count")+
  geom_vline(xintercept = 2000, linetype = "dashed")

ggsave(p, file=paste0(path, "Read_depth.tiff"))

p <- ggplot()+
  geom_histogram(mapping=aes(sample_sums(ps)), fill = "lightskyblue1", color="black", bins = 100)+
  theme_classic()+xlab("Read depth")+ylab("Count")+
  geom_vline(xintercept = 2000, linetype = "dashed")+xlim(c(0,100000))

ggsave(p, file=paste0(path, "Read_depth_zoomed.tiff"))

sum(sample_sums(ps)< 2000) #264 samples
sum(sample_sums(ps)< 2000)/4005 #Remove 6.6% of data

sample_data(ps)$N <- sample_sums(ps)

#Min read. depth = 2000:
ps <- prune_samples(sample_sums(ps)>=2000, ps) #3741 samples

saveRDS(ps, paste0(path, "ps_count_sub.rds"))

#Transform into relab and save:
ps.prop <- transform_sample_counts(ps, function(otu) otu/sum(otu))
saveRDS(ps.prop, paste0(path, "ps_relab_sub.rds"))
```

## Calculation of beta diversity
```{r cache = TRUE}
#Bray Curtis: 
beta_div_relab_bray = phyloseq::distance(ps.prop, method='bray')
beta_div_relab_bray %>% as.vector %>% summary

#Jaccard:
beta_div_relab_jaccard = phyloseq::distance(ps.prop, method='jaccard', binary = TRUE)
beta_div_relab_jaccard %>% as.vector %>% summary

saveRDS(beta_div_relab_bray, paste0(path, "BrayCurtis.rds"))
saveRDS(beta_div_relab_jaccard, paste0(path, "Jaccard.rds"))
```

## Visualize beta diversity in NMDS plots
```{r cache = TRUE}
#Bray Curtis
# k=2, stress = 0.2273045
set.seed(0)
BRAY_NMDS_relab=metaMDS(beta_div_relab_bray, k=2,trymax=30)
BRAY_NMDS_relab$stress
data.scores = as.data.frame(scores(BRAY_NMDS_relab))

data.scores$project <- sample_data(ps.prop)$BioProject
data.scores$Instrument <- sample_data(ps.prop)$Instrument
data.scores$Disease <- sample_data(ps.prop)$Disease

write.csv(data.scores, paste0(path, "beta_plot_BrayCurtis.txt"), quote = F)

data.scores <- read.csv(paste0(path, "Data/beta_plot_BrayCurtis.txt"))

#Plots:
ggplot(data=data.scores, aes(x=NMDS1, y=NMDS2, color=project))+
  geom_point(size=1)+theme_classic()

ggplot(data=data.scores, aes(x=NMDS1, y=NMDS2, color=Instrument))+
  geom_point(size=1)+theme_classic()

ggplot(data=data.scores, aes(x=NMDS1, y=NMDS2, color=Disease))+
  geom_point(size=1)+theme_classic()

fig_1a <-  ggplot(data=data.scores, aes(x=NMDS1, y=NMDS2, color=project))+
  geom_point(size=1)+theme_classic()+
  theme(text = element_text(family="sans", size=8), axis.title = element_text(family="sans", size=8), axis.text = element_text(family="sans", size=8), legend.position = "none") 

#Jaccard:
set.seed(0)
JAC_NMDS_relab=metaMDS(beta_div_relab_jaccard,k=2,trymax=30)
JAC_NMDS_relab$stress
# k=2, stress = 0.2007563

data.scores = as.data.frame(scores(JAC_NMDS_relab))

data.scores$project <- sample_data(ps.prop)$BioProject
data.scores$Instrument <- sample_data(ps.prop)$Instrument
data.scores$Disease <- sample_data(ps.prop)$Disease

write.csv(data.scores, paste0(path, "beta_plot_Jaccard.txt"), quote = F)

data.scores <- read.csv(paste0(path, "Data/beta_plot_Jaccard.txt"))


#Plots:
ggplot(data=data.scores, aes(x=NMDS1, y=NMDS2, color=project))+
  geom_point(size=1)+theme_classic()

ggplot(data=data.scores, aes(x=NMDS1, y=NMDS2, color=Instrument))+
  geom_point(size=1)+theme_classic()

ggplot(data=data.scores, aes(x=NMDS1, y=NMDS2, color=Disease))+
  geom_point(size=1)+theme_classic()

fig_1b <-  ggplot(data=data.scores, aes(x=NMDS1, y=NMDS2, color=project))+
  geom_point(size=1)+theme_classic()+
  theme(text = element_text(family="sans", size=8), axis.title = element_text(family="sans", size=8), axis.text = element_text(family="sans", size=8), legend.position = "none") 

fig_1ab <- ggarrange(fig_1a, fig_1b, labels = c("A)", "B)"), font.label = list(size=10, family="sans"))

```

Keep only ASV's present in at least 5% of the samples before calculating beta-div:
To reduce batch effect
```{r}
ps.prop <- readRDS(paste0(path, "ps_relab_sub.rds")) #3741 samples, 153773 taxa
ps.prop <- subset_samples(ps.prop, rowSums(otu_table(ps.prop))>0)

ps.prop # 3741 samples, 153773 taxa

genus <- as.data.frame(as.matrix(otu_table(ps.prop)))
genus[1:5, 1:5]
#Remove columns with no count in any sample:
genus1 <- genus[,colSums(genus)!=0] #151558 taxa

ps.prop <- phyloseq(otu_table(genus1, taxa_are_rows=F), sample_data(ps.prop), tax_table(ps.prop))
saveRDS(ps.prop, paste0(path, "ps_relab_sub_sub.rds"))

#Remove columns, where the genus is present in less than 2-5% of the samples:
sum(sapply(genus1, function(x) sum(x==0)) / dim(genus1)[1]<0.95) #399 taxa
sum(sapply(genus1, function(x) sum(x==0)) / dim(genus1)[1]<0.98) #1087 taxa
genus <- genus1[, sapply(genus1, function(x) sum(x==0)) / dim(genus1)[1]<0.95] 

ps_relab_reduced <- phyloseq(otu_table(genus, taxa_are_rows=F), sample_data(ps.prop), tax_table(ps.prop))

sum(rowSums(otu_table(ps_relab_reduced))==0)/3741 #0.276 -> loose 28% of samples
ps_relab_reduced <- subset_samples(ps_relab_reduced, rowSums(otu_table(ps_relab_reduced))>0)
saveRDS(ps_relab_reduced, paste0(path, "ps_relab_sub_5%.rds")) #399 taxa, 2708 samples

genus <- genus1[, sapply(genus1, function(x) sum(x==0)) / dim(genus1)[1]<0.98] 
ps_relab_reduced <- phyloseq(otu_table(genus, taxa_are_rows=F), sample_data(ps.prop), tax_table(ps.prop))
sum(rowSums(otu_table(ps_relab_reduced))==0)/3741 #0.135 --> loose 14% of samples
ps_relab_reduced <- subset_samples(ps_relab_reduced, rowSums(otu_table(ps_relab_reduced))>0)
saveRDS(ps_relab_reduced, paste0(path, "ps_relab_sub_2%.rds")) #1087 taxa, 3236 samples



#Make NMDS
#Bray Curtis: 
ps_relab_reduced<- readRDS(paste0(path, "ps_relab_sub_5%.rds"))

beta_div_relab_bray = phyloseq::distance(ps_relab_reduced, method='bray')
beta_div_relab_bray %>% as.vector %>% summary

#Jaccard:
beta_div_relab_jaccard = phyloseq::distance(ps_relab_reduced, method='jaccard', binary = TRUE)
beta_div_relab_jaccard %>% as.vector %>% summary

saveRDS(beta_div_relab_bray, paste0(path, "BrayCurtis_5%.rds"))
saveRDS(beta_div_relab_jaccard, paste0(path, "Jaccard_5%.rds"))

#NMDS
set.seed(0)
BRAY_NMDS_relab=metaMDS(beta_div_relab_bray, k=2,trymax=30)
BRAY_NMDS_relab$stress #0.2532493
data.scores = as.data.frame(scores(BRAY_NMDS_relab))

data.scores$project <- sample_data(ps_relab_reduced)$BioProject
data.scores$Instrument <- sample_data(ps_relab_reduced)$Instrument
data.scores$Disease <- sample_data(ps_relab_reduced)$Disease

write.csv(data.scores, paste0(path, "beta_plot_BrayCurtis_5%.txt"), quote = F)

data.scores <- read.csv(paste0(path, "Data/beta_plot_BrayCurtis_5%.txt"))

#Plots:
ggplot(data=data.scores, aes(x=NMDS1, y=NMDS2, color=project))+
  geom_point(size=1)+theme_classic()

ggplot(data=data.scores, aes(x=NMDS1, y=NMDS2, color=Instrument))+
  geom_point(size=1)+theme_classic()

ggplot(data=data.scores, aes(x=NMDS1, y=NMDS2, color=Disease))+
  geom_point(size=1)+theme_classic()

fig_1c <-  ggplot(data=data.scores, aes(x=NMDS1, y=NMDS2, color=project))+
  geom_point(size=1)+theme_classic()+
  theme(text = element_text(family="sans", size=8), axis.title = element_text(family="sans", size=8), axis.text = element_text(family="sans", size=8), legend.position = "none") 


set.seed(0)
JAC_NMDS_relab=metaMDS(beta_div_relab_jaccard,k=2,trymax=30)
JAC_NMDS_relab$stress # 0.1973994

data.scores = as.data.frame(scores(JAC_NMDS_relab))

data.scores$project <- sample_data(ps_relab_reduced)$BioProject
data.scores$Instrument <- sample_data(ps_relab_reduced)$Instrument
data.scores$Disease <- sample_data(ps_relab_reduced)$Disease

write.csv(data.scores, paste0(path, "beta_plot_Jaccard_5%.txt"), quote = F)

data.scores <- read.csv(paste0(path, "Data/beta_plot_Jaccard_5%.txt"))

#Plots:
ggplot(data=data.scores, aes(x=NMDS1, y=NMDS2, color=project))+
  geom_point(size=1)+theme_classic()

ggplot(data=data.scores, aes(x=NMDS1, y=NMDS2, color=Instrument))+
  geom_point(size=1)+theme_classic()

ggplot(data=data.scores, aes(x=NMDS1, y=NMDS2, color=Disease))+
  geom_point(size=1)+theme_classic()

fig_1d <-  ggplot(data=data.scores, aes(x=NMDS1, y=NMDS2, color=project))+
  geom_point(size=1)+theme_classic()+
  theme(text = element_text(family="sans", size=8), axis.title = element_text(family="sans", size=8), axis.text = element_text(family="sans", size=8), legend.position = "none") 

fig_1cd <- ggarrange(fig_1c, fig_1d, labels = c("C)", "C)"))
fig_1 <- ggarrange(fig_1ab, fig_1cd, nrow=2, ncol=1)
ggsave(fig_1, filename=paste0(path, "Illustrations/NMDS_QC.tiff"), width = 174, height = 174, units="mm", dpi=300)

```

Make 5% counts ps:
```{r}
ps <- readRDS(paste0(path, "ps_count.rds"))
ps_relab <- readRDS(paste0(path, "ps_relab_sub_5%.rds"))

my_subset <- subset(otu_table(ps), rownames(otu_table(ps)) %in% rownames(otu_table(ps_relab)))
new_ps <- merge_phyloseq(my_subset, tax_table(ps), sample_data(ps))
new_ps <- prune_taxa(row.names(tax_table(ps_relab)), new_ps) #399 taxa, 2708 samples

saveRDS(new_ps, paste0(path, "ps_count_5%.rds"))
```

## Overview of samples
```{r}
ps_count <- readRDS(paste0(path,"Data/ps_relab_sub_5%.rds"))
df <- as.data.frame(as.matrix(sample_data(ps_count)))

#Check stats:
length(unique(df$BioProject)) #47 -> 46 excluding asthma
unique(df$BioProject)
table(df$Disease, useNA = "always")
unique(df$Continent) #5
unique(df$Country) #19

df %>% dplyr::group_by(Disease) %>% summarise(N=n(), mean_age=mean(as.numeric(Age)), projects = length(unique(BioProject)), continents = length(unique(Continent)))

#CD
df_CD <- df %>% filter(Disease =="CD")
unique(df_CD$BioProject)
unique(df_CD$Continent)
df_sub <- df %>% filter(BioProject %in% unique(df_CD$BioProject))
table(df_sub$Disease)

#CeD
df_CeD <- df %>% filter(Disease =="CeD")
unique(df_CeD$BioProject)
unique(df_CeD$Continent)
df_sub <- df %>% filter(BioProject %in% unique(df_CeD$BioProject))
table(df_sub$Disease)

#GD
df_GD <- df %>% filter(Disease =="GD")
unique(df_GD$BioProject)
unique(df_GD$Continent)
df_sub <- df %>% filter(BioProject %in% unique(df_GD$BioProject))
table(df_sub$Disease)

#MG
df_MG <- df %>% filter(Disease =="MG")
unique(df_MG$BioProject)
unique(df_MG$Continent)
df_sub <- df %>% filter(BioProject %in% unique(df_MG$BioProject))
table(df_sub$Disease)

#MS
df_MS <- df %>% filter(Disease =="MS")
unique(df_MS$BioProject)
unique(df_MS$Continent)
df_sub <- df %>% filter(BioProject %in% unique(df_MS$BioProject))
table(df_sub$Disease)

#Ps
df_Ps <- df %>% filter(Disease =="Psoriasis")
unique(df_Ps$BioProject)
unique(df_Ps$Continent)
df_sub <- df %>% filter(BioProject %in% unique(df_Ps$BioProject))
table(df_sub$Disease)

#RA
df_RA <- df %>% filter(Disease =="RA")
unique(df_RA$BioProject)
unique(df_RA$Continent)
df_sub <- df %>% filter(BioProject %in% unique(df_RA$BioProject))
table(df_sub$Disease)

#SLE
df_SLE <- df %>% filter(Disease =="SLE")
unique(df_SLE$BioProject)
unique(df_SLE$Continent)
df_sub <- df %>% filter(BioProject %in% unique(df_SLE$BioProject))
table(df_sub$Disease)

#T1D
df_T1D <- df %>% filter(Disease =="T1D")
unique(df_T1D$BioProject)
unique(df_T1D$Continent)
df_sub <- df %>% filter(BioProject %in% unique(df_T1D$BioProject))
table(df_sub$Disease)

#UC
df_UC <- df %>% filter(Disease =="UC")
unique(df_UC$BioProject)
unique(df_UC$Continent)
df_sub <- df %>% filter(BioProject %in% unique(df_UC$BioProject))
table(df_sub$Disease)


#Make plots:
df_projects <- df %>% dplyr::group_by(Instrument) %>% summarise(samples = n(), projects = length(unique(BioProject)))

p1 <- ggplot(data=df_projects)+
  geom_bar(aes(x=reorder(Instrument,-projects, sum), y=projects,), stat="identity")+xlab("Instrument")+ylab("Studies")+theme_classic()+ theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

p2 <- ggplot(data=df_projects)+
  geom_bar(aes(x=reorder(Instrument,-samples), y=samples), stat="identity")+xlab("Instrument")+ylab("Total samples")+theme_classic()+ theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))


df_disease <- df %>% dplyr::group_by(Disease, sex) %>% summarise(samples = n())
df_disease$Disease <- ifelse(df_disease$Disease == "Psoriasis", "Ps", df_disease$Disease)

p3 <- ggplot(data=df_disease %>% filter(Disease!="asthma"))+
  geom_bar(aes(x=reorder(Disease,-samples, sum), y=samples, fill=sex), stat="identity")+xlab("Disease status")+ylab("Total samples")+theme_classic()+labs(fill="Sex")

ggarrange(p1, p3, labels=c("A", "B"), widths = c(0.5,1))
ggarrange(p2, p3, labels=c("A", "B"))



df_disease <- df_disease%>% filter(Disease!="asthma") %>% mutate(disease_full = ifelse(Disease=="CeD", "Celiac disease", ifelse(Disease=="GD", "Graves' disease", ifelse(Disease=="MS", "Multiple sclerosis", ifelse(Disease=="Ps", "Psoriasis", ifelse(Disease=="RA", "Rheumatoid arthritis", ifelse(Disease=="MG", "Myasthenia gravis", ifelse(Disease=="SLE", "Systemic lupus erythematosus", ifelse(Disease=="T1D", "Type 1 diabetes", ifelse(Disease=="CD", "Crohn's disease", ifelse(Disease=="UC", "Ulcarative colitis", "Healthy controls")))))))))))

p3 <- ggplot(data=df_disease)+
  geom_bar(aes(x=reorder(disease_full,-samples, sum), y=samples, fill=sex), stat="identity")+xlab("")+ylab("Total samples")+theme_classic()+labs(fill="Sex")+ theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

ggsave(p3, file = paste0(path, "/Illustrations/Disease_sex_age_avail.tiff"), width = 150, height = 120, units = "mm", dpi = 300, bg="white")

```
