==> Microbiota/4_Diversity_analyses.Rmd <==
---
title: "Diversity analyses"
output:
  github_document:
    toc: true
editor_options:
  chunk_output_type: console
---

## Read in packages and data
```{r message=FALSE, warning=FALSE, error=TRUE}
library(dplyr)
library(tidyverse)
library(ggplot2)
library(phyloseq)
library(vegan)
library(ggpubr)
library(car)

packageVersion("dplyr")
packageVersion("tidyverse")
packageVersion("ggplot2")
packageVersion("phyloseq")
packageVersion("vegan")
packageVersion("ggpubr")
packageVersion("car")

path <- "C:/Users/FX76TZ/OneDrive - Aalborg Universitet/Disease correlations/"
#path <- "~/vol1/all/"
```


## Beta diversity:

Read in data:
```{r}
ps_relab <- readRDS(paste0(path, "Data/ps_relab_sub_5%.rds"))

#Meta data:
df1 <- data.frame(sample_data(ps_relab))
df1$N <- as.numeric(df1$N)
```

Analysis:
```{r}
set.seed(1)
#Bray Curtis:
res1 <- adonis(otu_table(ps_relab)~ N + Instrument+Continent + Disease, data=df1, method = "bray")
write.table(res1, file = paste0(path, "BrayCurtis_res.txt"))

set.seed(1)
#Jaccard:
res2 <- adonis2(otu_table(ps_relab)~ N + Instrument+Continent + Disease, data=df1, method="jaccard", binary=T)
write.table(res2, file = paste0(path, "Jaccard_res.txt"))

#Including age and sex
ps_sub <- subset_samples(ps_relab, !is.na(Age))
ps_sub <- subset_samples(ps_sub, !is.na(sex))
df2 <- data.frame(sample_data(ps_sub))
df2$N <- as.numeric(df2$N)
df2$Age <- as.numeric(df2$Age)

set.seed(1)
#Bray Curtis:
res3 <- adonis(otu_table(ps_sub)~ N + Instrument+Continent + Age + sex + Disease, data=df2, method = "bray")
write.table(res3, file = paste0(path, "BrayCurtis_res_SexAge.txt")) #OBS:HVORDAN GEMME??

set.seed(1)
#Jaccard:
res4 <- adonis(otu_table(ps_sub)~ N + Instrument+Continent + Age + sex + Disease, data=df2, method="jaccard", binary=T)
write.table(res4, file = paste0(path, "Jaccard_res_SexAge.txt"))
```

Remove data:
```{r}
rm(ps_relab, df2)
```
