==> Microbiota/5_Microbiota_disease_profile_correlation.Rmd <==
---
title: "Statistical analysis: Single Genus"
output:
  github_document:
    toc: true
editor_options:
  chunk_output_type: console
---


## Read in packages and data
```{r message=FALSE, warning=FALSE, error=TRUE}
library(tidyverse)
library(ggplot2)
library(dplyr)
library(phyloseq)
library(ggpubr)
library(forestploter)
library(grid)

packageVersion("tidyverse")
packageVersion("ggplot2")
packageVersion("dplyr")
packageVersion("phyloseq")
packageVersion("ggpubr")

library(Maaslin2)
packageVersion("Maaslin2")

path <- "C:/Users/FX76TZ/OneDrive - Aalborg Universitet/Disease correlations/"

predict_col <- c("#4F2641", "#EAE4E3", "#FEF2FA", "#FADBC4", "#FFEAD7", "#FEF5ED",  "#C2C6B5",  "#F7FBED", "#F8F9F3", "#898A87", "#000000", "#E0E1DC", "#FFFFFF")
accent_col <- c("#CD68AB", "#85416E", "#FF906D", "#FFC498", "#A5BC64", "#B1CD64","#E0F3A6")

hc_col <- accent_col[6]
cd1_col <- accent_col[1]
cd2_col <- accent_col[2]
uc1_col <- accent_col[4]
uc2_col <- accent_col[3]

```

Load data:
```{r}
ps_relab <- readRDS(paste0(path, "Data/ps_relab_sub_5%.rds")) #399 taxa, 2708 samples
```

## Prepare data for analysis: 
Select only samples with sex and age information. Keep all HC, but subset cases to have same number of patients of each disease.
```{r}
ps_relab <- subset_samples(ps_relab, !is.na(Age)) #858 samples
ps_relab <- subset_samples(ps_relab, !is.na(sex)) #782 samples
table(sample_data(ps_relab)$Disease)

sample_ps <- function(ps, FUN = sample, ...){
  ids <- sample_names(ps)
  sampled_ids <- FUN(ids, ...)
  ps <- prune_samples(sampled_ids, ps)
  return(ps)
}
ps_CD <- subset_samples(ps_relab, Disease == "CD")
ps_CeD <- subset_samples(ps_relab, Disease == "CeD")
ps_GD <- subset_samples(ps_relab, Disease == "GD")
ps_MS <- subset_samples(ps_relab, Disease == "MS")
ps_Ps <- subset_samples(ps_relab, Disease == "Psoriasis")
ps_RA <- subset_samples(ps_relab, Disease == "RA")
ps_UC <- subset_samples(ps_relab, Disease == "UC")
ps_HC <- subset_samples(ps_relab, Disease == "HC")


```


### MaAsLin2:
```{r}
#For CD-UC: downsample CD to 37

ps_sub <- subset_samples(ps_relab, Disease %in% c("HC", "UC"))

input_data <- as.data.frame(as.matrix(otu_table(ps_sub)))
input_metadata <- as.data.frame(as.matrix(sample_data(ps_sub)))
input_metadata$Age <- as.numeric(input_metadata$Age)

setwd(paste0(path, "Results/"))
fit_data = Maaslin2(
    input_data = input_data, 
    input_metadata = input_metadata, 
    output = "UC_n37", 
    max_significance = 1,
    min_prevalence = 0,
    transform = "AST",
    normalization = "None",
    fixed_effects = c("Disease","Instrument", "Continent", "Age", "sex"), reference = c("Disease,HC", "Instrument,Illumina MiSeq", "Continent,Asia"),
    plot_heatmap = F, plot_scatter = F)

res <- read.csv("UC_n37/all_results.tsv", sep="\t")
res <- res %>% filter(metadata == "Disease")
head(res)

write.table(res, paste0(path, "Results/UC_n37.txt"), col.names=T, row.names=F)

set.seed(1)
ps_sub <- sample_ps(ps_CD, size=37)
ps_sub <- subset_samples(ps_relab, Run %in% c(sample_data(ps_sub)$Run, sample_data(ps_HC)$Run))

input_data <- as.data.frame(as.matrix(otu_table(ps_sub)))
input_metadata <- as.data.frame(as.matrix(sample_data(ps_sub)))
input_metadata$Age <- as.numeric(input_metadata$Age)

setwd(paste0(path, "Results/"))
fit_data = Maaslin2(
    input_data = input_data, 
    input_metadata = input_metadata, 
    output = "CD_n37", 
    max_significance = 1,
    min_prevalence = 0,
    transform = "AST",
    normalization = "None",
    fixed_effects = c("Disease","Instrument", "Continent", "Age", "sex"), reference = c("Disease,HC", "Instrument,Illumina MiSeq", "Continent,Asia"),
    plot_heatmap = F, plot_scatter = F)

res <- read.csv("CD_n37/all_results.tsv", sep="\t")
res <- res %>% filter(metadata == "Disease")
head(res)

write.table(res, paste0(path, "Results/CD_n37.txt"), col.names=T, row.names=F)

```

Correlation with CeD: Downsample CD and UC to 30
```{r}
ps_sub <- subset_samples(ps_relab, Disease %in% c("HC", "CeD"))

input_data <- as.data.frame(as.matrix(otu_table(ps_sub)))
input_metadata <- as.data.frame(as.matrix(sample_data(ps_sub)))
input_metadata$Age <- as.numeric(input_metadata$Age)

setwd(paste0(path, "Results/"))
fit_data = Maaslin2(
    input_data = input_data, 
    input_metadata = input_metadata, 
    output = "CeD_n30", 
    max_significance = 1,
    min_prevalence = 0,
    transform = "AST",
    normalization = "None",
    fixed_effects = c("Disease","Instrument", "Continent", "Age", "sex"), reference = c("Disease,HC", "Instrument,Illumina MiSeq", "Continent,Asia"),
    plot_heatmap = F, plot_scatter = F)

res <- read.csv("CeD_n30/all_results.tsv", sep="\t")
res <- res %>% filter(metadata == "Disease")
head(res)

write.table(res, paste0(path, "Results/CeD_n30.txt"), col.names=T, row.names=F)

set.seed(1)
ps_sub <- sample_ps(ps_CD, size=30)
ps_sub <- subset_samples(ps_relab, Run %in% c(sample_data(ps_sub)$Run, sample_data(ps_HC)$Run))

input_data <- as.data.frame(as.matrix(otu_table(ps_sub)))
input_metadata <- as.data.frame(as.matrix(sample_data(ps_sub)))
input_metadata$Age <- as.numeric(input_metadata$Age)

setwd(paste0(path, "Results/"))
fit_data = Maaslin2(
    input_data = input_data, 
    input_metadata = input_metadata, 
    output = "CD_n30", 
    max_significance = 1,
    min_prevalence = 0,
    transform = "AST",
    normalization = "None",
    fixed_effects = c("Disease","Instrument", "Continent", "Age", "sex"), reference = c("Disease,HC", "Instrument,Illumina MiSeq", "Continent,Asia"),
    plot_heatmap = F, plot_scatter = F)

res <- read.csv("CD_n30/all_results.tsv", sep="\t")
res <- res %>% filter(metadata == "Disease")
head(res)

write.table(res, paste0(path, "Results/CD_n30.txt"), col.names=T, row.names=F)

set.seed(1)
ps_sub <- sample_ps(ps_UC, size=30)
ps_sub <- subset_samples(ps_relab, Run %in% c(sample_data(ps_sub)$Run, sample_data(ps_HC)$Run))

input_data <- as.data.frame(as.matrix(otu_table(ps_sub)))
input_metadata <- as.data.frame(as.matrix(sample_data(ps_sub)))
input_metadata$Age <- as.numeric(input_metadata$Age)

setwd(paste0(path, "Results/"))
fit_data = Maaslin2(
    input_data = input_data, 
    input_metadata = input_metadata, 
    output = "UC_n30", 
    max_significance = 1,
    min_prevalence = 0,
    transform = "AST",
    normalization = "None",
    fixed_effects = c("Disease","Instrument", "Continent", "Age", "sex"), reference = c("Disease,HC", "Instrument,Illumina MiSeq", "Continent,Asia"),
    plot_heatmap = F, plot_scatter = F)

res <- read.csv("UC_n30/all_results.tsv", sep="\t")
res <- res %>% filter(metadata == "Disease")
head(res)

write.table(res, paste0(path, "Results/UC_n30.txt"), col.names=T, row.names=F)
```

Correlation with GD: Downsample GD to 37 and 47. And Run CD 47
```{r}
ps_sub <- subset_samples(ps_relab, Disease %in% c("HC", "CD"))

input_data <- as.data.frame(as.matrix(otu_table(ps_sub)))
input_metadata <- as.data.frame(as.matrix(sample_data(ps_sub)))
input_metadata$Age <- as.numeric(input_metadata$Age)

setwd(paste0(path, "Results/"))
fit_data = Maaslin2(
    input_data = input_data, 
    input_metadata = input_metadata, 
    output = "CD_n47", 
    max_significance = 1,
    min_prevalence = 0,
    transform = "AST",
    normalization = "None",
    fixed_effects = c("Disease","Instrument", "Continent", "Age", "sex"), reference = c("Disease,HC", "Instrument,Illumina MiSeq", "Continent,Asia"),
    plot_heatmap = F, plot_scatter = F)

res <- read.csv("CD_n47/all_results.tsv", sep="\t")
res <- res %>% filter(metadata == "Disease")
head(res)

write.table(res, paste0(path, "Results/CD_n47.txt"), col.names=T, row.names=F)

set.seed(1)
ps_sub <- sample_ps(ps_GD, size=47)
ps_sub <- subset_samples(ps_relab, Run %in% c(sample_data(ps_sub)$Run, sample_data(ps_HC)$Run))

input_data <- as.data.frame(as.matrix(otu_table(ps_sub)))
input_metadata <- as.data.frame(as.matrix(sample_data(ps_sub)))
input_metadata$Age <- as.numeric(input_metadata$Age)

setwd(paste0(path, "Results/"))
fit_data = Maaslin2(
    input_data = input_data, 
    input_metadata = input_metadata, 
    output = "GD_n47", 
    max_significance = 1,
    min_prevalence = 0,
    transform = "AST",
    normalization = "None",
    fixed_effects = c("Disease","Instrument", "Continent", "Age", "sex"), reference = c("Disease,HC", "Instrument,Illumina MiSeq", "Continent,Asia"),
    plot_heatmap = F, plot_scatter = F)

res <- read.csv("GD_n47/all_results.tsv", sep="\t")
res <- res %>% filter(metadata == "Disease")
head(res)

write.table(res, paste0(path, "Results/GD_n47.txt"), col.names=T, row.names=F)

set.seed(1)
ps_sub <- sample_ps(ps_GD, size=37)
ps_sub <- subset_samples(ps_relab, Run %in% c(sample_data(ps_sub)$Run, sample_data(ps_HC)$Run))

input_data <- as.data.frame(as.matrix(otu_table(ps_sub)))
input_metadata <- as.data.frame(as.matrix(sample_data(ps_sub)))
input_metadata$Age <- as.numeric(input_metadata$Age)

setwd(paste0(path, "Results/"))
fit_data = Maaslin2(
    input_data = input_data, 
    input_metadata = input_metadata, 
    output = "GD_n37", 
    max_significance = 1,
    min_prevalence = 0,
    transform = "AST",
    normalization = "None",
    fixed_effects = c("Disease","Instrument", "Continent", "Age", "sex"), reference = c("Disease,HC", "Instrument,Illumina MiSeq", "Continent,Asia"),
    plot_heatmap = F, plot_scatter = F)

res <- read.csv("GD_n37/all_results.tsv", sep="\t")
res <- res %>% filter(metadata == "Disease")
head(res)

write.table(res, paste0(path, "Results/GD_n37.txt"), col.names=T, row.names=F)
```

Correlation with MS: Downsample CD and UC to 35
```{r}
ps_sub <- subset_samples(ps_relab, Disease %in% c("HC", "MS"))

input_data <- as.data.frame(as.matrix(otu_table(ps_sub)))
input_metadata <- as.data.frame(as.matrix(sample_data(ps_sub)))
input_metadata$Age <- as.numeric(input_metadata$Age)

setwd(paste0(path, "Results/"))
fit_data = Maaslin2(
    input_data = input_data, 
    input_metadata = input_metadata, 
    output = "MS_n35", 
    max_significance = 1,
    min_prevalence = 0,
    transform = "AST",
    normalization = "None",
    fixed_effects = c("Disease","Instrument", "Continent", "Age", "sex"), reference = c("Disease,HC", "Instrument,Illumina MiSeq", "Continent,Asia"),
    plot_heatmap = F, plot_scatter = F)

res <- read.csv("MS_n35/all_results.tsv", sep="\t")
res <- res %>% filter(metadata == "Disease")
head(res)

write.table(res, paste0(path, "Results/MS_n35.txt"), col.names=T, row.names=F)

set.seed(1)
ps_sub <- sample_ps(ps_CD, size=35)
ps_sub <- subset_samples(ps_relab, Run %in% c(sample_data(ps_sub)$Run, sample_data(ps_HC)$Run))

input_data <- as.data.frame(as.matrix(otu_table(ps_sub)))
input_metadata <- as.data.frame(as.matrix(sample_data(ps_sub)))
input_metadata$Age <- as.numeric(input_metadata$Age)

setwd(paste0(path, "Results/"))
fit_data = Maaslin2(
    input_data = input_data, 
    input_metadata = input_metadata, 
    output = "CD_n35", 
    max_significance = 1,
    min_prevalence = 0,
    transform = "AST",
    normalization = "None",
    fixed_effects = c("Disease","Instrument", "Continent", "Age", "sex"), reference = c("Disease,HC", "Instrument,Illumina MiSeq", "Continent,Asia"),
    plot_heatmap = F, plot_scatter = F)

res <- read.csv("CD_n35/all_results.tsv", sep="\t")
res <- res %>% filter(metadata == "Disease")
head(res)

write.table(res, paste0(path, "Results/CD_n35.txt"), col.names=T, row.names=F)

set.seed(1)
ps_sub <- sample_ps(ps_UC, size=35)
ps_sub <- subset_samples(ps_relab, Run %in% c(sample_data(ps_sub)$Run, sample_data(ps_HC)$Run))

input_data <- as.data.frame(as.matrix(otu_table(ps_sub)))
input_metadata <- as.data.frame(as.matrix(sample_data(ps_sub)))
input_metadata$Age <- as.numeric(input_metadata$Age)

setwd(paste0(path, "Results/"))
fit_data = Maaslin2(
    input_data = input_data, 
    input_metadata = input_metadata, 
    output = "UC_n35", 
    max_significance = 1,
    min_prevalence = 0,
    transform = "AST",
    normalization = "None",
    fixed_effects = c("Disease","Instrument", "Continent", "Age", "sex"), reference = c("Disease,HC", "Instrument,Illumina MiSeq", "Continent,Asia"),
    plot_heatmap = F, plot_scatter = F)

res <- read.csv("UC_n35/all_results.tsv", sep="\t")
res <- res %>% filter(metadata == "Disease")
head(res)

write.table(res, paste0(path, "Results/UC_n35.txt"), col.names=T, row.names=F)
```

Correlation with Ps: Downsample Ps to 37 and 47
```{r}

set.seed(1)
ps_sub <- sample_ps(ps_Ps, size=47)
ps_sub <- subset_samples(ps_relab, Run %in% c(sample_data(ps_sub)$Run, sample_data(ps_HC)$Run))

input_data <- as.data.frame(as.matrix(otu_table(ps_sub)))
input_metadata <- as.data.frame(as.matrix(sample_data(ps_sub)))
input_metadata$Age <- as.numeric(input_metadata$Age)

setwd(paste0(path, "Results/"))
fit_data = Maaslin2(
    input_data = input_data, 
    input_metadata = input_metadata, 
    output = "Ps_n47", 
    max_significance = 1,
    min_prevalence = 0,
    transform = "AST",
    normalization = "None",
    fixed_effects = c("Disease","Instrument", "Continent", "Age", "sex"), reference = c("Disease,HC", "Instrument,Illumina MiSeq", "Continent,Asia"),
    plot_heatmap = F, plot_scatter = F)

res <- read.csv("Ps_n47/all_results.tsv", sep="\t")
res <- res %>% filter(metadata == "Disease")
head(res)

write.table(res, paste0(path, "Results/Ps_n47.txt"), col.names=T, row.names=F)

set.seed(1)
ps_sub <- sample_ps(ps_Ps, size=37)
ps_sub <- subset_samples(ps_relab, Run %in% c(sample_data(ps_sub)$Run, sample_data(ps_HC)$Run))

input_data <- as.data.frame(as.matrix(otu_table(ps_sub)))
input_metadata <- as.data.frame(as.matrix(sample_data(ps_sub)))
input_metadata$Age <- as.numeric(input_metadata$Age)

setwd(paste0(path, "Results/"))
fit_data = Maaslin2(
    input_data = input_data, 
    input_metadata = input_metadata, 
    output = "Ps_n37", 
    max_significance = 1,
    min_prevalence = 0,
    transform = "AST",
    normalization = "None",
    fixed_effects = c("Disease","Instrument", "Continent", "Age", "sex"), reference = c("Disease,HC", "Instrument,Illumina MiSeq", "Continent,Asia"),
    plot_heatmap = F, plot_scatter = F)

res <- read.csv("Ps_n37/all_results.tsv", sep="\t")
res <- res %>% filter(metadata == "Disease")
head(res)

write.table(res, paste0(path, "Results/Ps_n37.txt"), col.names=T, row.names=F)
```

Correlation with RA: Downsample RA to 37 and 47
```{r}

set.seed(1)
ps_sub <- sample_ps(ps_RA, size=47)
ps_sub <- subset_samples(ps_relab, Run %in% c(sample_data(ps_sub)$Run, sample_data(ps_HC)$Run))

input_data <- as.data.frame(as.matrix(otu_table(ps_sub)))
input_metadata <- as.data.frame(as.matrix(sample_data(ps_sub)))
input_metadata$Age <- as.numeric(input_metadata$Age)

setwd(paste0(path, "Results/"))
fit_data = Maaslin2(
    input_data = input_data, 
    input_metadata = input_metadata, 
    output = "RA_n47", 
    max_significance = 1,
    min_prevalence = 0,
    transform = "AST",
    normalization = "None",
    fixed_effects = c("Disease","Instrument", "Continent", "Age", "sex"), reference = c("Disease,HC", "Instrument,Illumina MiSeq", "Continent,Asia"),
    plot_heatmap = F, plot_scatter = F)

res <- read.csv("RA_n47/all_results.tsv", sep="\t")
res <- res %>% filter(metadata == "Disease")
head(res)

write.table(res, paste0(path, "Results/RA_n47.txt"), col.names=T, row.names=F)

set.seed(1)
ps_sub <- sample_ps(ps_RA, size=37)
ps_sub <- subset_samples(ps_relab, Run %in% c(sample_data(ps_sub)$Run, sample_data(ps_HC)$Run))

input_data <- as.data.frame(as.matrix(otu_table(ps_sub)))
input_metadata <- as.data.frame(as.matrix(sample_data(ps_sub)))
input_metadata$Age <- as.numeric(input_metadata$Age)

setwd(paste0(path, "Results/"))
fit_data = Maaslin2(
    input_data = input_data, 
    input_metadata = input_metadata, 
    output = "RA_n37", 
    max_significance = 1,
    min_prevalence = 0,
    transform = "AST",
    normalization = "None",
    fixed_effects = c("Disease","Instrument", "Continent", "Age", "sex"), reference = c("Disease,HC", "Instrument,Illumina MiSeq", "Continent,Asia"),
    plot_heatmap = F, plot_scatter = F)

res <- read.csv("RA_n37/all_results.tsv", sep="\t")
res <- res %>% filter(metadata == "Disease")
head(res)

write.table(res, paste0(path, "Results/RA_n37.txt"), col.names=T, row.names=F)
```


Make correlation:
```{r}

#CD-UC:
res1 <- read.csv(paste0(path, "Results/CD_n37.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
res2 <- read.csv(paste0(path, "Results/UC_n37.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
t <- cor.test(res1$t, res2$t, method = "pearson", alternative = "two.sided")

cor_df <- data.frame(disease1="CD", disease2="UC", cor=t$estimate, ci1=t$conf.int[1], ci2=t$conf.int[2], p=t$p.value, n=37)

#CeD:
res1 <- read.csv(paste0(path, "Results/CD_n30.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
res2 <- read.csv(paste0(path, "Results/CeD_n30.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
t <- cor.test(res1$t, res2$t, method = "pearson", alternative = "two.sided")

cor_df <- rbind(cor_df, data.frame(disease1="CD", disease2="CeD", cor=t$estimate, ci1=t$conf.int[1], ci2=t$conf.int[2], p=t$p.value, n=30))

res1 <- read.csv(paste0(path, "Results/UC_n30.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
t <- cor.test(res1$t, res2$t, method = "pearson", alternative = "two.sided")

cor_df <- rbind(cor_df, data.frame(disease1="UC", disease2="CeD", cor=t$estimate, ci1=t$conf.int[1], ci2=t$conf.int[2], p=t$p.value, n=30))

#GD:
res1 <- read.csv(paste0(path, "Results/CD_n47.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
res2 <- read.csv(paste0(path, "Results/GD_n47.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
t <- cor.test(res1$t, res2$t, method = "pearson", alternative = "two.sided")

cor_df <- rbind(cor_df, data.frame(disease1="CD", disease2="GD", cor=t$estimate, ci1=t$conf.int[1], ci2=t$conf.int[2], p=t$p.value, n=47))

res1 <- read.csv(paste0(path, "Results/UC_n37.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
res2 <- read.csv(paste0(path, "Results/GD_n37.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
t <- cor.test(res1$t, res2$t, method = "pearson", alternative = "two.sided")

cor_df <- rbind(cor_df, data.frame(disease1="UC", disease2="GD", cor=t$estimate, ci1=t$conf.int[1], ci2=t$conf.int[2], p=t$p.value, n=37))

#MS:
res1 <- read.csv(paste0(path, "Results/CD_n35.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
res2 <- read.csv(paste0(path, "Results/MS_n35.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
t <- cor.test(res1$t, res2$t, method = "pearson", alternative = "two.sided")

cor_df <- rbind(cor_df, data.frame(disease1="CD", disease2="MS", cor=t$estimate, ci1=t$conf.int[1], ci2=t$conf.int[2], p=t$p.value, n=35))

res1 <- read.csv(paste0(path, "Results/UC_n35.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
t <- cor.test(res1$t, res2$t, method = "pearson", alternative = "two.sided")

cor_df <- rbind(cor_df, data.frame(disease1="UC", disease2="MS", cor=t$estimate, ci1=t$conf.int[1], ci2=t$conf.int[2], p=t$p.value, n=35))

#Ps:
res1 <- read.csv(paste0(path, "Results/CD_n47.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
res2 <- read.csv(paste0(path, "Results/Ps_n47.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
t <- cor.test(res1$t, res2$t, method = "pearson", alternative = "two.sided")

cor_df <- rbind(cor_df, data.frame(disease1="CD", disease2="Ps", cor=t$estimate, ci1=t$conf.int[1], ci2=t$conf.int[2], p=t$p.value, n=47))

res1 <- read.csv(paste0(path, "Results/UC_n37.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
res2 <- read.csv(paste0(path, "Results/Ps_n37.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
t <- cor.test(res1$t, res2$t, method = "pearson", alternative = "two.sided")

cor_df <- rbind(cor_df, data.frame(disease1="UC", disease2="Ps", cor=t$estimate, ci1=t$conf.int[1], ci2=t$conf.int[2], p=t$p.value, n=37))

#RA:
res1 <- read.csv(paste0(path, "Results/CD_n47.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
res2 <- read.csv(paste0(path, "Results/RA_n47.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
t <- cor.test(res1$t, res2$t, method = "pearson", alternative = "two.sided")

cor_df <- rbind(cor_df, data.frame(disease1="CD", disease2="RA", cor=t$estimate, ci1=t$conf.int[1], ci2=t$conf.int[2], p=t$p.value, n=47))

res1 <- read.csv(paste0(path, "Results/UC_n37.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
res2 <- read.csv(paste0(path, "Results/RA_n37.txt"), sep=" ") %>% mutate(t=coef/stderr) %>% arrange(feature)
t <- cor.test(res1$t, res2$t, method = "pearson", alternative = "two.sided")

cor_df <- rbind(cor_df, data.frame(disease1="UC", disease2="RA", cor=t$estimate, ci1=t$conf.int[1], ci2=t$conf.int[2], p=t$p.value, n=37))

write.table(cor_df, paste0(path, "Results/Correlation_16S_new.txt"), sep="\t", quote=F, row.names = F, col.names = T)

```

Correlation analysis:
```{r}

df_plot <- read.csv(paste0(path, "Results/Correlation_16S_new.txt"), sep="\t")

df_plot$p_adj <- p.adjust(df_plot$p, method = "BH")

#update naming:
df_plot <- df_plot %>% mutate(disease_full = ifelse(disease2=="CeD", "Celiac disease", ifelse(disease2=="GD", "Graves' disease", ifelse(disease2=="MS", "Multiple sclerosis", ifelse(disease2=="Ps", "Psoriasis", ifelse(disease2=="RA", "Rheumatoid arthritis", "Ulcerative colitis"))))))

disease_order <- c("Ulcerative colitis", "Graves' disease", "Multiple sclerosis", "Celiac disease", "Psoriasis", "Rheumatoid arthritis")

df_plot$disease_full <- factor(df_plot$disease_full, level=disease_order)
df_plot <- df_plot %>% arrange(disease_full, disease1)

df_plot$plot.column <- "                                         "
df_plot$"ρ [95% CI]" <- paste0(formatC(df_plot$cor, 2, format = "f"), " [", formatC(df_plot$ci1, 2, format = "f"), "-", formatC(df_plot$ci2, 2, format = "f"), "]")

df_plot$P <- ifelse(df_plot$p_adj<0.01, formatC(df_plot$p_adj, format="e", digits=2), formatC(df_plot$p_adj, format="f", digits=3))
df_plot$P  <- gsub("e", "%*%10^", df_plot$P)

dt_fig <- df_plot %>% mutate(disease_full=as.character(disease_full)) %>% select(disease_full, disease1, n, plot.column, "ρ [95% CI]", P)  %>% rename("  " = disease1, " " = plot.column, Disease=disease_full, "P adjusted"=P, N=n)
dt_fig[c(seq(3,11,2)),1] <- ""


parse_mat <- matrix(FALSE, 
                    nrow = nrow(df_plot),
                    ncol = 6)
parse_mat[,6] <- TRUE 

fig1 <- forest(data=dt_fig, est=df_plot$cor, lower=df_plot$ci1, upper=df_plot$ci2, ci_column = 4, sizes=0.5, theme=forest_theme(base_size = 8, base_family = "sans", core=list(fg_params = list(parse=parse_mat))), xlab="ρ [95% CI]", xlim=c(-1,1))

fig1 <- edit_plot(fig1, 
               row = c(1,seq(2,11,2)), col=4,
               which = c("ci"), 
               gp = gpar(col = cd2_col))

fig1 <- edit_plot(fig1, 
               row = c(seq(3,11,2)), col=4,
               which = c("ci"), 
               gp = gpar(col = uc2_col))

fig1 <- edit_plot(fig1, 
               row = c(1,4,5,8,9), 
               which = "background", 
               gp = gpar(fill = predict_col[13]))

fig1 <- edit_plot(fig1, 
               row = c(2,3,6,7,10,11), 
               which = "background", 
               gp = gpar(fill = predict_col[12], alpha=0.5))


ggsave(fig1, file = paste0(path, "/Illustrations/Microbiota_correlation.tiff"), width = 150, height = 100, units = "mm", dpi = 300, bg="white")

```
